통합. read에 있던 코드 일단 삭제.
