+++ Sumdays/app/src/main/res/layout/activity_daily_write.xml
@@ -1,19 +1,105 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:padding="16dp"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    tools:context=".DailyWriteActivity">
 
-    <TextView
-        android:id="@+id/textDiary"
-        android:textSize="18sp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/memo_list_view"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/input_area"
+        android:layout_marginStart="32dp"
+        android:layout_marginEnd="32dp"
+        android:layout_marginTop="32dp"
+        android:padding="8dp"
+        android:background="@drawable/border_dark_grey"
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager" />
 
-    <Button
-        android:id="@+id/btnEdit"
-        android:text="수정하기"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"/>
-</LinearLayout>
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/input_area"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toBottomOf="@id/memo_list_view"
+        app:layout_constraintBottom_toTopOf="@id/sum_button"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginStart="48dp"
+        android:layout_marginEnd="48dp"
+        android:layout_marginTop="16dp"
+        android:layout_marginBottom="16dp"
+        android:background="@drawable/border_dark_grey"> <EditText
+        android:id="@+id/memo_input_edittext"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:hint="하루를 기록해보세요"
+        android:background="@drawable/rounded_edittext_background"
+        android:padding="12dp"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toStartOf="@id/mic_icon"
+        android:layout_marginEnd="8dp" />
+
+        <ImageView
+            android:id="@+id/mic_icon"
+            android:layout_width="32dp"
+            android:layout_height="32dp"
+            android:src="@drawable/ic_mic"
+            app:layout_constraintTop_toTopOf="@id/memo_input_edittext"
+            app:layout_constraintBottom_toBottomOf="@id/memo_input_edittext"
+            app:layout_constraintEnd_toStartOf="@id/photo_icon"
+            android:layout_marginEnd="8dp"
+            android:contentDescription="음성 입력" />
+
+        <ImageView
+            android:id="@+id/photo_icon"
+            android:layout_width="32dp"
+            android:layout_height="32dp"
+            android:src="@drawable/ic_image"
+            app:layout_constraintTop_toTopOf="@id/memo_input_edittext"
+            app:layout_constraintBottom_toBottomOf="@id/memo_input_edittext"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:contentDescription="사진 입력" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+    <ImageView
+        android:id="@+id/sum_button"
+        android:layout_width="48dp"
+        android:layout_height="48dp"
+        android:src="@drawable/sum_fox_button"
+        app:layout_constraintTop_toBottomOf="@id/input_area"
+        app:layout_constraintBottom_toTopOf="@id/sum_button_divider"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:contentDescription="SUM 버튼"
+        android:layout_marginTop="8dp"
+        android:layout_marginBottom="8dp"
+        android:background="@drawable/border_dark_grey" />
+
+    <View
+    android:id="@+id/sum_button_divider"
+    android:layout_width="match_parent"
+    android:layout_height="2dp"
+    android:background="#A0A0A0"
+    app:layout_constraintTop_toBottomOf="@id/sum_button"
+    app:layout_constraintBottom_toTopOf="@id/bottom_buttons_layout"
+    android:layout_marginStart="32dp"
+    android:layout_marginEnd="32dp"
+    android:layout_marginBottom="16dp"
+    android:backgroundTint="#606060" />
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+    android:id="@+id/bottom_buttons_layout"
+    android:layout_width="match_parent"
+    android:layout_height="64dp"
+    app:layout_constraintTop_toBottomOf="@id/sum_button_divider"
+    app:layout_constraintBottom_toBottomOf="parent"
+    android:paddingHorizontal="32dp"
+    android:background="@drawable/border_dark_grey"> </androidx.constraintlayout.widget.ConstraintLayout>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
